extract debian package for hddl_reset

+++ hddl/Ubuntu/hddl_container/benchmark/image_build.sh
@@ -6,7 +6,7 @@ export HDDL_RESOURCE_FOLDER=/home/chunfung/CF/Benchmark_WS
 export HOST_PACKAGE_LINK=https://ubit-artifactory-sh.intel.com/artifactory/sed-dgn-local/yocto/builds/2020/PREINT/20201003-0611
 export HOST_PACKAGE_NAME=bypass_host_hddlunite_hvasample_demo-2327426.tgz
 export OPENVINO_PACKAGE_NAME=l_openvino_toolkit_private_ubuntu18_kmb_x86_p_2021.1.0-1237-3cabe58ed07-releases_2020_kmb_pv2.tar.gz
-
+export DEB_PACKAGE_NAME=kmb-hddl-driver-dkms_0.1.0-eabaa_all.deb
 
 if [ -z "$HDDL_IMAGE_NAME" ]; then
     echo "Need to set HDDL_IMAGE_NAME "
@@ -38,6 +38,11 @@ if [ -z "$OPENVINO_PACKAGE_NAME" ]; then
     exit
 fi
 
+if [ -z "$DEB_PACKAGE_NAME" ]; then
+    echo "Need to set DEB_PACKAGE_NAME "
+    exit
+fi
+
 mkdir -p $HDDL_RESOURCE_FOLDER
 #TODO
 rm -rf $HDDL_RESOURCE_FOLDER/$HOST_PACKAGE_NAME
@@ -46,6 +51,9 @@ wget $HOST_PACKAGE_LINK/host_packages/$HOST_PACKAGE_NAME -P $HDDL_RESOURCE_FOLDE
 rm -rf $HDDL_RESOURCE_FOLDER/$OPENVINO_PACKAGE_NAME
 wget $HOST_PACKAGE_LINK/$OPENVINO_PACKAGE_NAME -P $HDDL_RESOURCE_FOLDER
 
+rm -rf $HDDL_RESOURCE_FOLDER/$DEB_PACKAGE_NAME
+wget $HOST_PACKAGE_LINK/$DEB_PACKAGE_NAME -P $HDDL_RESOURCE_FOLDER
+
 cp -r ./scripts/* $HDDL_RESOURCE_FOLDER
 
 docker build \
@@ -53,6 +61,7 @@ docker build \
 -f ./dockerfiletest \
 --build-arg HOST_PACKAGE_NAME \
 --build-arg OPENVINO_PACKAGE_NAME \
+--build-arg DEB_PACKAGE_NAME \
 -t $HDDL_IMAGE_NAME:$HDDL_IMAGE_TAG $HDDL_RESOURCE_FOLDER
 
 docker image ls
